+++ components/about/TechStackCell.tsx
