+++ components/projects/ProjectCard.tsx
@@ -16,23 +16,24 @@ import { useEffect, useState } from "react";
 
 export default function ProjectCard({ projectData }: { projectData: Project }) {
   const { title, name, description, imageUrl, techStack } = projectData;
-  const { enroute } = useAppContext();
+  const { router, enroute } = useAppContext();
   const [isDesktop, setIsDesktop] = useState(true);
+  const path = `/projects/${name}`;
 
   useEffect(() => {
     if (
       /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(
-        navigator.userAgent
+        navigator.userAgent,
       )
     ) {
       setIsDesktop(false);
     } else {
       setIsDesktop(true);
     }
+    router?.prefetch(path);
   }, []);
 
-  function onClickHandler(name: string) {
-    const path = `/projects/${name}`;
+  function onClickHandler() {
     enroute(path, { scroll: true });
   }
 
@@ -99,7 +100,7 @@ export default function ProjectCard({ projectData }: { projectData: Project }) {
         width="full"
         height="full"
         rounded="inherit"
-        onClick={() => onClickHandler(name)}
+        onClick={() => onClickHandler()}
       >
         <Image
           src={imageUrl}
