+++ package.json
@@ -17,6 +17,7 @@
     "@fontsource/montserrat": "^5.0.8",
     "@tsparticles/engine": "^3.2.1",
     "@tsparticles/react": "^3.0.0",
+    "@tsparticles/slim": "^3.2.1",
     "@types/node": "^20.11.16",
     "@types/react": "^18.2.53",
     "@types/react-dom": "^18.2.18",
@@ -32,7 +33,6 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "styled-components": "^6.1.8",
-    "tsparticles": "^3.2.1",
     "typescript": "^5.3.3"
   }
 }
